final fix for the frontend

+++ 4.6.0/src/MellowoodMedical.Application/CMS/Dtos/InsertOrUpdateCMSInput.cs
@@ -8,7 +8,7 @@ namespace MellowoodMedical.CMS.Dtos
 	public class InsertOrUpdateCMSInput
 	{
 		[Required]
-		public int PageId { get; set; }
+		public long PageId { get; set; }
 
 		public string PageName { get; set; }
